Fix yet another int that hsould be an ssize_t. (dm)
git-svn-id: svn://mielke.cc/main/brltty@6261 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/usb_linux.c
@@ -686,7 +686,7 @@ usbWriteEndpoint (
         struct usbdevfs_urb *urb = usbInterruptTransfer(endpoint, (void *)buffer, length, timeout);
 
         if (urb) {
-          int count = urb->actual_length;
+          ssize_t count = urb->actual_length;
           free(urb);
           return count;
         }
